Fixed error in 'batch size' input from cli

+++ cli.py
@@ -35,7 +35,7 @@ def app():
     name = "Image_1.png"
 
     if args_0 in sys.argv: in_kaggle = True
-    if args_1 in sys.argv: batch_size = float(sys.argv[sys.argv.index(args_1) + 1])
+    if args_1 in sys.argv: batch_size = int(sys.argv[sys.argv.index(args_1) + 1])
     if args_2 in sys.argv: lr = float(sys.argv[sys.argv.index(args_2) + 1])
     if args_3 in sys.argv: wd = float(sys.argv[sys.argv.index(args_3) + 1])
     if args_4 in sys.argv:
